# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BigPecas Backend API
  version: 1.0.0
  description: |
    API REST do BigPecas para autenticacao, catalogo, pecas, pedidos, frete,
    wishlist, notificacoes, avaliacoes e administracao.

    A maioria das rotas exige token Supabase Auth no header
    `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:3001
    description: Ambiente local
...
security:
  - bearerAuth: []
paths:
...
  /api/pedidos:
    get:
      tags: [Pedidos]
      summary: Lista compras do usuario
      description: Endpoint de compatibilidade que retorna somente compras do usuario autenticado.
      responses:
        "200":
          description: Compras do usuario.
          content:
            application/json:
              example:
                - id: "2026-123456"
                  user_id: 12
                  status: pago
                  total: 462.46
                  visao: compra
    post:
      tags: [Pedidos]
      summary: Cria pedido
      description: Cria pedido com itens, frete, endereco e forma de pagamento; tambem gera codigo de rastreio e vendas por fornecedor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoCreateRequest"
            example:
              itens:
                - id: 45
                  quantidade: 1
              frete:
                transportadora: Loggi
                tipo: Loggi Ponto
                valor: 18.46
                prazo_texto: Entrega em até 6 dias úteis
              cupom: null
              endereco:
                nome: Francisco Martins
                cep: 80240-140
                logradouro: Rua 12
                numero: "12"
                complemento: "1"
                bairro: Centro
                cidade: Curitiba
                uf: PR
                telefone: "41902111119"
              forma_pagamento:
                id: pix
                nome: PIX
                descricao: Aprovação imediata
      responses:
        "201":
          description: Pedido criado.
          content:
            application/json:
              example:
                id: "2026-123456"
                status: aguardando_pagamento
                total: 462.46
                codigo_rastreio: BG123456789AB
                visao: compra
...
  /api/pedidos/{id}:
    get:
      tags: [Pedidos]
      summary: Busca pedido por id
      description: Retorna a visao de compra ou venda conforme usuario e query `visao=venda`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: "2026-123456"
        - name: visao
          in: query
          schema: { type: string, enum: [compra, venda, vendas] }
      responses:
        "200":
          description: Pedido encontrado.
          content:
            application/json:
              example:
                id: "2026-123456"
                status: pago
                itens:
                  - id: 45
                    nome: Carburador Solex H40
                    quantidade: 1
                visao: compra
  /api/pedidos/{id}/status:
    patch:
      tags: [Pedidos]
      summary: Atualiza status do pedido
      description: |
        Altera o status respeitando transicoes validas:
        `aguardando_pagamento -> pago/cancelado`, `pago -> enviado/cancelado`,
        `enviado -> entregue`.

        Ao mudar status, o backend sincroniza vendas e registra/enviar notificacoes.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: "2026-123456"
      requestBody:
        required: true
        content:
          application/json:
            example:
              status: enviado
      responses:
        "200":
          description: Status atualizado.
          content:
            application/json:
              example:
                id: "2026-123456"
                status: enviado
                historico:
                  - status: aguardando_pagamento
                    data: "2026-08-24T13:00:00.000Z"
                  - status: pago
                    data: "2026-08-24T13:05:00.000Z"
                  - status: enviado
                    data: "2026-08-24T14:00:00.000Z"
...
  /api/avaliacoes/pedidos/{pedidoId}:
    get:
      tags: [Avaliacoes]
      summary: Consulta estado de avaliacoes do pedido
      description: Retorna se a avaliacao esta liberada e quais fornecedores/produtos ja foram avaliados.
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Estado de avaliacoes.
          content:
            application/json:
              example:
                pedido_id: "2026-123456"
                status: entregue
                liberada: true
                fornecedores:
                  - fornecedor_id: 12
                    fornecedor_nome: Auto Pecas Martins
                    avaliado: false
                produtos:
                  - peca_id: 45
                    nome: Carburador Solex H40
                    avaliado: false
...
  /api/admin/pedidos:
    get:
      tags: [Admin]
      summary: Lista pedidos para administracao
      description: Lista pedidos com paginacao e filtro opcional por status.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          $ref: "#/components/responses/AdminList"
  /api/admin/pedidos/{id}/status:
    patch:
      tags: [Admin]
      summary: Corrige status de pedido como administrador
      description: Atualiza status de pedido e registra no historico que a alteracao foi feita por administrador.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            example:
              status: cancelado
      responses:
        "200":
          description: Status atualizado.
          content:
            application/json:
              example:
                message: Status atualizado pelo administrador.
                pedido:
                  id: "2026-123456"
                  status: cancelado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
...
    AdminList:
      description: Lista administrativa paginada.
      content:
        application/json:
          example:
            data:
              - id: 12
                email: francisco.martins@example.com
            pagination:
              page: 1
              limit: 20
              total: 1
  schemas:
...
    OrderStatus:
      type: string
      enum:
        - aguardando_pagamento
        - pago
        - enviado
        - entregue
        - cancelado
    PedidoCreateRequest:
      type: object
      required: [itens, endereco, forma_pagamento]
      properties:
        itens:
          type: array
          items:
            type: object
            required: [id, quantidade]
            properties:
              id: { type: integer }
              quantidade: { type: integer }
        frete:
          type: object
        cupom:
          nullable: true
        endereco:
          type: object
          required: [nome, cep, logradouro, numero, bairro, cidade, uf, telefone]
          properties:
            nome: { type: string }
            cep: { type: string }
            logradouro: { type: string }
            numero: { type: string }
            complemento: { type: string }
            bairro: { type: string }
            cidade: { type: string }
            uf: { type: string }
            telefone: { type: string }
        forma_pagamento:
          type: object